pack: use a doubly linked lru list for eviction
Instead of going over every possible position in the hash table, use a
list to keep track of lru base, so we already know which ones to prefer.

Go with what git does as well, and limit the cache to 256 elements
instead of using a memory size limit.

This does make it perform better, but only by a smaller margin that I
had hoped.