Add filter flag to prevent preloading workdir data
This bumps the custom filter structure definition and adds flags
to the new version.

This also adds support for a flag that says that a custom filter
can support directly reading working directory file content instead
of having the libgit2 filter framework preread the file data.  This
allows a filter to skip loading the on-disk data entirely into
memory when going in the TO_ODB direction (provided that the filter
is applied first the chain of filters).