transport: Lock usage of custom_transports
It was previously stated in the documentation that callers of
`git_transport_register` are responsible for synchronizing with other calls as
well as "other calls to the library that instantiate transports". When writing a
binding library for libgit2, however, this is quite difficult as the list of
APIs that need to be synchronized are:

* git_transport_register
* git_transport_unregister
* git_remote_connect
* git_push_finish
* git_remote_fetch
* git_remote_upload
* git_remote_push
* git_clone
* ... (I may have missed some)

This list is also growing over time, so it's quite difficult to ensure that a
library is properly synchronized in a general fashion.

This commit adds a global `git_mutex` which protects access to the global
`custom_transports` list as well as mentioning in the documentation that
transports created by the callback may linger even after the transport is
unregistered.