Race condition in fs refs iterator when refs are packed by another process.
When refs are packed, a new pack file is created containing all previously
packed refs as well as any newly packed loose refs. The existing pack file is
then replaced by the new pack file, and any included loose refs are deleted.

In order to avoid a race condition when another process is packing refs, the
loose refs must be loaded before the pack file is loaded.

The bug can be demonstrated in the current code by placing a sleep after the
call to packed_reload and packing the refs in another process. As a result,
any refs that were transitioning from loose to packed will be omited from the
list of refs returned by the iterator.