Fix leaking of the TLS objects on POSIX platforms
Previously we only free the global state for the thread
on which the actual shutdown is performed. Global state
objects belonging to other threads actually got leaked.
This patch fixes the issue by freeing the thread specific
global object at the beginning of git_libgit2_shutdown
regardless the number of the init. This way we require
that each thread call git_libgit2_shutdown for proper
cleanup, which is reasonable.