Fix leaking of the TLS objects on POSIX platforms
Previously we free the global state for the thread on
which the actual shutdown is performed and then delete
the TLS key. Global state objects belonging to other
threads actually got leaked if those threads haven't
exited when the actual shutdown happens. This is
because the TLS key got deleted and as a result the
destructor cb__free_status won't be called for those
threads. This patch fixes the issue by keeping the
TLS key as long as possible. Although with this patch
the TLS key will never get deleted, it's OK because only
one key is globally alive until the whole process exits.